pyproject.toml
[project]
name = "stryde-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.3.0",
    "cryptography>=44.0.0",
    "fastapi[standard]>=0.129.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=1.4",
    "uvicorn[standard]>=0.41.0",
]
